Move default similarity metrics names to SimilarityMetrics
