+++ xquant/keras/keras_report_utils.py
@@ -1,18 +1,18 @@
 #  Copyright 2024 Sony Semiconductor Israel, Inc. All rights reserved.
-#  #
+#
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
 #  You may obtain a copy of the License at
-#  #
+#
 #      http://www.apache.org/licenses/LICENSE-2.0
-#  #
+#
 #  Unless required by applicable law or agreed to in writing, software
 #  distributed under the License is distributed on an "AS IS" BASIS,
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 #  ==============================================================================
-#
+
 import logging
 
 import keras
@@ -29,7 +29,7 @@ from functools import partial
 
 from xquant import XQuantConfig
 from xquant.common.constants import INTERMEDIATE_METRICS_REPR, INTERMEDIATE_METRICS_VAL, XQUANT_REPR, XQUANT_VAL
-from xquant.common.framework_report_utils import FrameworkReportUtils, MSE_METRIC_NAME, CS_METRIC_NAME, SQNR_METRIC_NAME
+from xquant.common.framework_report_utils import FrameworkReportUtils
 from model_compression_toolkit.ptq.keras.quantization_facade import DEFAULT_KERAS_TPC
 from model_compression_toolkit.core.keras.reader.reader import model_reader
 from xquant.common.model_folding_utils import ModelFoldingUtils
